refactor: rename routes config file
