+++ _codux/board-wrappers/component-wrapper.tsx
@@ -1,7 +1,7 @@
 import { createRemixStub } from '@remix-run/testing';
 import { PropsWithChildren } from 'react';
 import { EcomAPIContextProvider } from '~/api/ecom-api-context-provider';
-import { ROUTES } from '~/router/config';
+import { ROUTES } from '~/routes';
 
 export interface ComponentWrapperProps extends PropsWithChildren {
     loaderData?: Record<string, unknown>;
