+++ app/routes/_index/route.tsx
@@ -1,9 +1,9 @@
 import { LinksFunction, LoaderFunctionArgs } from '@remix-run/node';
-import { Link, MetaFunction, useLoaderData, useNavigate, json } from '@remix-run/react';
+import { json, Link, MetaFunction, useLoaderData, useNavigate } from '@remix-run/react';
 import { getEcomApi } from '~/api/ecom-api';
 import { HeroImage } from '~/components/hero-image/hero-image';
 import { ProductCard } from '~/components/product-card/product-card';
-import { ROUTES } from '~/router/config';
+import { ROUTES } from '~/routes';
 import { getUrlOriginWithPath, isOutOfStock } from '~/utils';
 import styles from './index.module.scss';
