+++ src/routes.ts
@@ -6,6 +6,11 @@ const CATEGORY = '/category/:categorySlug';
 const PRODUCT = `/products/:productSlug`;
 const THANK_YOU = '/thank-you';
 
+/**
+ * This object centralizes link generation to maintain simplicity and consistency.
+ * Note: Remix uses a file-based routing system, so modifying this object will not
+ * affect the actual routes or pages available in the application.
+ */
 export const ROUTES = {
     home: { path: HOME, to: () => HOME },
     about: { path: ABOUT, to: () => ABOUT },
@@ -18,11 +23,4 @@ export const ROUTES = {
         path: PRODUCT,
         to: (slug: string) => generatePath(PRODUCT, { productSlug: slug }),
     },
-};
-
-export type ROUTE_KEYS = keyof typeof ROUTES;
-
-export type RouteParams = {
-    [PRODUCT]: { slug: string };
-    [CATEGORY]: { name: string };
 };
